refctor: 케어콜 시간 설정 컨트롤러, request 스키마 변경
