+++ src/main/java/com/example/medicare_call/controller/action/ElderHealthInfoController.java
@@ -14,23 +14,15 @@ import org.springframework.web.bind.annotation.*;
 
 @Tag(name = "ElderHealthInfo", description = "어르신 건강정보(질환, 복약주기, 특이사항) 등록 API")
 @RestController
-@RequestMapping("/elders/{elderId}")
+@RequestMapping("/elders")
 @RequiredArgsConstructor
 public class ElderHealthInfoController {
     private final ElderHealthInfoService elderHealthInfoService;
-    private final CareCallSettingService careCallSettingService;
 
     @Operation(summary = "어르신 건강정보 등록", description = "질환, 복약주기, 특이사항을 등록합니다.")
-    @PostMapping("/health-info")
+    @PostMapping("/{elderId}/health-info")
     public ResponseEntity<Void> registerElderHealthInfo(@PathVariable Integer elderId, @Valid @RequestBody ElderHealthRegisterRequest request) {
         elderHealthInfoService.registerElderHealthInfo(elderId, request);
         return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @Operation(summary = "어르신 전화 시간대 등록", description = "3번의 케어콜 시간대를 저장합니다.")
-    @PostMapping("/care-call-setting")
-    public ResponseEntity<Void> settingCarCallInfo(@PathVariable Integer elderId, @Valid @RequestBody CareCallSettingRequest request) {
-        careCallSettingService.settingCareCall(elderId,request);
-        return ResponseEntity.status(HttpStatus.OK).build();
     }
 }
