+++ src/main/java/com/example/medicare_call/service/CareCallService.java
@@ -7,6 +7,7 @@ import com.example.medicare_call.domain.MedicationSchedule;
 import com.example.medicare_call.global.enums.CallType;
 import com.example.medicare_call.repository.*;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -20,6 +21,7 @@ import java.util.List;
 import java.util.Map;
 
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class CareCallService {
@@ -58,9 +60,11 @@ public class CareCallService {
             HttpHeaders headers = new HttpHeaders();
             headers.setContentType(MediaType.APPLICATION_JSON);
 
+            String normalizedPhoneNumber = normalizeKoreanPhoneNumber(phoneNumber);
+
             Map<String, Object> body = new HashMap<>();
             body.put("elderId", elderId);
-            body.put("phoneNumber", phoneNumber);
+            body.put("phoneNumber", normalizedPhoneNumber);
             body.put("prompt", prompt);
 
             HttpEntity<Map<String, Object>> request = new HttpEntity<>(body, headers);
@@ -70,5 +74,20 @@ public class CareCallService {
         } catch (Exception e) {
             System.err.println("호출 실패: " + e.getMessage());
         }
+    }
+
+    private String normalizeKoreanPhoneNumber(String phoneNumber) {
+        // 010으로 시작하는 경우 +82로 변환
+        if (phoneNumber.startsWith("010")) {
+            return "+82" + phoneNumber.substring(1);  // "01012345678" → "+821012345678"
+        }
+
+        // 이미 +82로 시작하면 그대로 반환
+        if (phoneNumber.startsWith("+82")) {
+            return phoneNumber;
+        }
+
+        log.warn("전화번호 형식이 예상과 다릅니다. 입력값: {}", phoneNumber);
+        return phoneNumber;
     }
 }
